dialog, filterchip, list, links converted to sb7
